PostProcessingController moved back to engine
