+++ app/src/GuiController.cpp
@@ -1,7 +1,9 @@
 #include <engine/graphics/GraphicsController.hpp>
-#include <imgui.h>
+#include <engine/graphics/PostProcessingController.hpp>
+
 #include <GuiController.hpp>
 #include <ProgramState.hpp>
+#include <imgui.h>
 
 namespace app {
     void GuiController::initialize() {
@@ -46,8 +48,8 @@ namespace app {
             ImGui::SliderFloat("Green", &Settings::getInstance().lightColor.y, 0.0f, 12.0f);
             ImGui::SliderFloat("Blue", &Settings::getInstance().lightColor.z, 0.0f, 12.0f);
         }
-
-        Filter filter = Settings::getInstance().filter;
+        auto ppc = engine::core::Controller::get<engine::graphics::PostProcessingController>();
+        Filter filter = ppc->get_active_filter();
         if (ImGui::CollapsingHeader("Cool Filters")) {
             ImGui::Text("Choose your filter:");
             if (ImGui::RadioButton("NONE", filter == Filter::NONE)) { filter = Filter::NONE; }
@@ -57,7 +59,7 @@ namespace app {
             if (ImGui::RadioButton("LINE", filter == Filter::OUTLINE)) { filter = Filter::OUTLINE; }
             if (ImGui::RadioButton("B&W", filter == Filter::BLACKWHITE)) { filter = Filter::BLACKWHITE; }
         }
-        Settings::getInstance().filter = filter;
+        ppc->set_active_filter(filter);
 
         ImGui::Spacing();
         ImGui::Separator();
